Start of bundle consumption
